// packages/capturer-extension/src/enums/messageTypes.ts
enum messageTypes {
  RunningStatus = "runningStatus",
  Log = "log",
  Start = "start",
  End = "end",
  PopupPageStart = "popupPageStart",
  PopupPageEnd = "popupPageEnd",
  PopupPageGetStatus = "popupPageGetStatus",
  PopupPageChangeStatus = "popupPageChangeStatus",
  PopupPageChangeStart = "popupPageChangeStart",
  PopupPageChangeEnd = "popupPageChangeEnd",
}

export default messageTypes;